# swagger-config.json
{
  "openapi": "3.0.0",
  "paths": {
    "/users": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      },
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
    "/roles": {
      "post": {
        "operationId": "RolesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRolDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Roles"
        ]
      },
      "get": {
        "operationId": "RolesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Roles"
        ]
      }
    },
    "/roles/{id}": {
      "get": {
        "operationId": "RolesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Roles"
        ]
      },
      "patch": {
        "operationId": "RolesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRolDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Roles"
        ]
      },
      "delete": {
        "operationId": "RolesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Roles"
        ]
      }
    },
    "/lenguajes": {
      "post": {
        "operationId": "LenguajesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLenguajeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Lenguajes"
        ]
      },
      "get": {
        "operationId": "LenguajesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Lenguajes"
        ]
      }
    },
    "/lenguajes/{id}": {
      "get": {
        "operationId": "LenguajesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Lenguajes"
        ]
      },
      "patch": {
        "operationId": "LenguajesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLenguajeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Lenguajes"
        ]
      }
    },
    "/lenguajes/{name}": {
      "delete": {
        "operationId": "LenguajesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Lenguajes"
        ]
      }
    },
    "/autos": {
      "post": {
        "operationId": "AutosController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAutoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Autos"
        ]
      },
      "get": {
        "operationId": "AutosController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Autos"
        ]
      }
    },
    "/autos/{id}": {
      "get": {
        "operationId": "AutosController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Autos"
        ]
      },
      "patch": {
        "operationId": "AutosController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAutoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Autos"
        ]
      }
    },
    "/autos/{name}": {
      "delete": {
        "operationId": "AutosController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Autos"
        ]
      }
    },
    "/direcciones": {
      "post": {
        "operationId": "DireccionesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDireccionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Direcciones"
        ]
      },
      "get": {
        "operationId": "DireccionesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Direcciones"
        ]
      }
    },
    "/direcciones/{id}": {
      "get": {
        "operationId": "DireccionesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Direcciones"
        ]
      },
      "patch": {
        "operationId": "DireccionesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDireccionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Direcciones"
        ]
      }
    },
    "/direcciones/{name}": {
      "delete": {
        "operationId": "DireccionesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Direcciones"
        ]
      }
    },
    "/login": {
      "post": {
        "operationId": "LoginController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Login"
        ]
      }
    }
  },
  "info": {
    "title": "Bootcamp",
    "description": "Bootcamp API",
    "version": "1.0.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "Direccion": {
        "type": "object",
        "properties": {}
      },
      "Rol": {
        "type": "object",
        "properties": {}
      },
      "Lenguaje": {
        "type": "object",
        "properties": {}
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string"
          },
          "apellido": {
            "type": "string"
          },
          "dni": {
            "type": "number"
          },
          "direccion": {
            "$ref": "#/components/schemas/Direccion"
          },
          "rol": {
            "$ref": "#/components/schemas/Rol"
          },
          "lenguaje": {
            "$ref": "#/components/schemas/Lenguaje"
          }
        },
        "required": [
          "nombre",
          "apellido",
          "dni",
          "direccion",
          "rol",
          "lenguaje"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {}
      },
      "CreateRolDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string"
          }
        },
        "required": [
          "nombre"
        ]
      },
      "UpdateRolDto": {
        "type": "object",
        "properties": {}
      },
      "CreateLenguajeDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string"
          }
        },
        "required": [
          "nombre"
        ]
      },
      "UpdateLenguajeDto": {
        "type": "object",
        "properties": {}
      },
      "User": {
        "type": "object",
        "properties": {}
      },
      "CreateAutoDto": {
        "type": "object",
        "properties": {
          "patente": {
            "type": "string"
          },
          "marca": {
            "type": "string"
          },
          "modelo": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": [
          "patente",
          "marca",
          "modelo",
          "color",
          "user"
        ]
      },
      "UpdateAutoDto": {
        "type": "object",
        "properties": {}
      },
      "CreateDireccionDto": {
        "type": "object",
        "properties": {
          "calle": {
            "type": "string"
          },
          "numero": {
            "type": "number"
          },
          "localidad": {
            "type": "string"
          },
          "provincia": {
            "type": "string"
          },
          "pais": {
            "type": "string"
          }
        },
        "required": [
          "calle",
          "numero",
          "localidad",
          "provincia",
          "pais"
        ]
      },
      "UpdateDireccionDto": {
        "type": "object",
        "properties": {}
      },
      "CreateLoginDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string"
          }
        },
        "required": [
          "nombre"
        ]
      }
    }
  }
}